+++ src/http_chunk.h
@@ -10,7 +10,7 @@ struct stat_cache_entry; /* declaration */
 
 int http_chunk_append_mem(request_st *r, const char * mem, size_t len); /* copies memory */
 int http_chunk_append_buffer(request_st *r, buffer *mem); /* may reset "mem" */
-int http_chunk_decode_append_mem(request_st * const r, const char * const mem, const size_t len);
+int http_chunk_decode_append_mem(request_st * const r, const char * const mem, const uint32_t len);
 int http_chunk_decode_append_buffer(request_st * const r, buffer * const mem); /* may reset "mem" */
 int http_chunk_transfer_cqlen(request_st *r, chunkqueue *src, size_t len);
 int http_chunk_append_file_fd(request_st *r, const buffer *fn, int fd, off_t sz);
